# MYTHICBEASTS: Performance improvement: Use the OAuth2 client credentials protocol