Motivation: Speed up integration tests.

I've tested this works.

I measured the speed up for requests (as measured between curl's `time_pretransfer` and `time_total`, i.e. ignoring DNS/handshake/auth):

* Before: 1s
* After: 0.6s

40% improvement for me. Further, since I'm located ~maximally from the API server (~350ms RTT), which accounts for most the 0.6s I observe, hopefully Github Actions isn't as far, and so observes an even larger proportional speedup.

Fixes https://github.com/StackExchange/dnscontrol/issues/3628